src/data/requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiohttp_socks==0.10.1
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.11
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
netaddr==1.3.0
passlib==1.7.4
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
python-socks==2.7.1
requests==2.32.3
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
WTForms==3.1.2
yarl==1.18.3
yookassa==3.5.0
